jFitEval can now be defined in problem.py
