+++ src/utils/utility.py
@@ -227,92 +227,6 @@ def collaborate(
     return complete_solutions_set_unique_typecasted
 
 
-def evaluate(
-        first_population, first_archive,
-        second_population, second_archive, joint_class, min_num_evals):
-    """Forms complete solutions, evaluates their joint fitness and evaluates
-    the individual fitness values.
-
-    :param first_population: the population (list) of scenarios.
-    :param first_archive: the archive (list) of scenarios.
-    :param second_population: the population of MLC outputs.
-    :param second_archive: the archive of MLC outputs.
-    :param joint_class: type into which each complete solution will
-                        be typecasted.
-    :param min_num_evals: the minimum number of collaborations and
-                          thus, joint fitness evaluations per
-                          individual.
-    :returns: set of complete solutions with their fitness values,
-              set of scenarios with their individual fitness values,
-              and the set of MLC outputs with their individual
-              fitness values.
-    """
-    # Exception handling must be added.
-
-    # Deep copy the inputs
-    # population_one = deepcopy(first_population)
-    # population_two = deepcopy(second_population)
-    # archive_one = deepcopy(first_archive)
-    # archive_two = deepcopy(second_archive)
-
-    first_component_class = type(first_population[0])
-    complete_solutions_set = collaborate(
-        first_archive,
-        first_population,
-        second_archive,
-        second_population,
-        joint_class,
-        first_component_class,
-        min_num_evals)
-
-    # Evaluate joint fitness and record its value.
-    for c in complete_solutions_set:
-        c.fitness.values = evaluate_joint_fitness(c)
-
-    # Evaluate individual fitness values.
-    for individual in first_population:
-        individual.fitness.values = evaluate_individual(
-            individual, complete_solutions_set, 0)
-
-    for individual in second_population:
-        individual.fitness.values = evaluate_individual(
-            individual, complete_solutions_set, 1)
-
-    return complete_solutions_set, first_population, second_population
-
-
-def evaluate_joint_fitness(c):
-    """Evaluates the joint fitness of a complete solution.
-
-    It takes the complete solution as input and returns its joint
-    fitness as output.
-    """
-    # # Returns a random value for now.
-    # return (random.uniform(-5.0, 5.0),)
-
-    # MTQ problem.
-    cf = 10  # Correction factor that controls the granularity of x and y.
-    x = c[0][0]
-    y = c[1][0]
-
-    h_1 = 50
-    x_1 = 0.75
-    y_1 = 0.75
-    s_1 = 1.6
-    f_1 = h_1 * \
-        (1 - ((16.0/s_1) * pow((x/cf - x_1), 2)) -
-         ((16.0/s_1) * pow((y/cf - y_1), 2)))
-    h_2 = 150
-    x_2 = 0.25
-    y_2 = 0.25
-    s_2 = 1.0/32.0
-    f_2 = h_2 * \
-        (1 - ((16.0/s_2) * pow((x/cf - x_2), 2)) -
-         ((16.0/s_2) * pow((y/cf - y_2), 2)))
-
-    return (max(f_1, f_2),)
-
-
 def evaluate_individual(individual, complete_solution_set, index):
     """Aggregates joint fitness values that an individual has been invovled in.
