+++ tests/test_evaluate.py
@@ -1,6 +1,6 @@
 from unittest.case import SkipTest, skip, skipIf
 from deap import creator
-from src.utils.utility import evaluate
+from src.main.ICCEA import ICCEA
 import problem
 import unittest
 
@@ -25,7 +25,7 @@ class TestEvaluate(unittest.TestCase):
                [[4, True, -7.8], [[1, 'a'], [21, 'd']]]]
 
         # FIXME: this function failed, possibly due to an error in evaluate_joint_fitness()
-        a, b, c = evaluate(pScen, aScen, pMLCO, aMLCO, cls, k)
+        a, b, c = self.evaluate(pScen, aScen, pMLCO, aMLCO, cls, k)
         self.assertEqual((a, b, c), (css, pScen, pMLCO))
 
         # Check whether fitness values for complete solutions and
